+++ backend/internal/eventlog/EventLog.go
@@ -14,7 +14,7 @@ type Event struct {
 	ID            string          `json:"id"`
 	Type          string          `json:"type"`
 	AccountID     string          `json:"account_id,omitempty"`
-	LogicalClock  int             `json:"logical_clock"`
+	LogicalClock  int64           `json:"logical_clock"`
 	UnixTimestamp int64           `json:"unix_timestamp"`
 	Data          json.RawMessage `json:"data"`
 }
@@ -26,7 +26,7 @@ type AppendInput struct {
 }
 
 type GetEventIteratorInput struct {
-	FromCursor int
+	FromCursor int64
 }
 
 type EventLog interface {
@@ -42,7 +42,7 @@ type EventIterator interface {
 
 type ProjectInput struct {
 	EventLog   EventLog
-	FromCursor int
+	FromCursor int64
 	Apply      func(ctx context.Context, event *Event)
 }
 
@@ -58,7 +58,7 @@ func Project(ctx context.Context, input *ProjectInput) {
 
 type StreamToChannelInput struct {
 	EventLog   EventLog
-	FromCursor int
+	FromCursor int64
 	C          chan *Event
 }
